# api/openapi/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiError": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"details": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryType": {
				"type": "string",
				"enum": [
					"bill",
					"income",
					"variable"
				]
			},
...
			"BillingPeriod": {
				"type": "string",
				"enum": [
					"monthly",
					"bi_weekly",
					"weekly",
					"semi_annually"
				]
			},
			"Income": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"billing_period": {
						"$ref": "#/components/schemas/BillingPeriod"
					},
					"start_date": {
						"type": "string"
					},
					"day_of_month": {
						"type": "number",
						"format": "double"
					},
					"recurrence_week": {
						"type": "number",
						"format": "double"
					},
					"recurrence_day": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"due_day": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"amount",
					"billing_period",
					"payment_source_id",
					"is_active",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateIncomeRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"billing_period": {
						"$ref": "#/components/schemas/BillingPeriod"
					},
					"start_date": {
						"type": "string"
					},
					"day_of_month": {
						"type": "number",
						"format": "double"
					},
					"recurrence_week": {
						"type": "number",
						"format": "double"
					},
					"recurrence_day": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"due_day": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"amount",
					"billing_period",
					"payment_source_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateIncomeRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"billing_period": {
						"$ref": "#/components/schemas/BillingPeriod"
					},
					"start_date": {
						"type": "string"
					},
					"day_of_month": {
						"type": "number",
						"format": "double"
					},
					"recurrence_week": {
						"type": "number",
						"format": "double"
					},
					"recurrence_day": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"due_day": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Category": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"is_predefined": {
						"type": "boolean"
					},
					"sort_order": {
						"type": "number",
						"format": "double"
					},
					"color": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/CategoryType"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"is_predefined",
					"sort_order",
					"color",
					"type",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/CategoryType"
					},
					"color": {
						"type": "string"
					},
					"sort_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCategoryRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/CategoryType"
					},
					"color": {
						"type": "string"
					},
					"sort_order": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReorderCategoriesRequest": {
				"properties": {
					"order": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Array of category IDs in desired order"
					}
				},
				"required": [
					"order"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Bill": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"billing_period": {
						"$ref": "#/components/schemas/BillingPeriod"
					},
					"start_date": {
						"type": "string"
					},
					"day_of_month": {
						"type": "number",
						"format": "double"
					},
					"recurrence_week": {
						"type": "number",
						"format": "double"
					},
					"recurrence_day": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"due_day": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"amount",
					"billing_period",
					"payment_source_id",
					"is_active",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBillRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"billing_period": {
						"$ref": "#/components/schemas/BillingPeriod"
					},
					"start_date": {
						"type": "string"
					},
					"day_of_month": {
						"type": "number",
						"format": "double"
					},
					"recurrence_week": {
						"type": "number",
						"format": "double"
					},
					"recurrence_day": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"due_day": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"amount",
					"billing_period",
					"payment_source_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBillRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"billing_period": {
						"$ref": "#/components/schemas/BillingPeriod"
					},
					"start_date": {
						"type": "string"
					},
					"day_of_month": {
						"type": "number",
						"format": "double"
					},
					"recurrence_week": {
						"type": "number",
						"format": "double"
					},
					"recurrence_day": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"due_day": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "BudgetForFun API",
		"version": "1.0.0",
		"description": "BudgetForFun API - Personal budget tracking application",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "BudgetForFun Team"
		}
	},
	"paths": {
...
		"/api/incomes": {
			"get": {
				"operationId": "GetIncomes",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Income"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "inc-123",
												"name": "Salary",
												"amount": 500000,
												"billing_period": "bi_weekly",
												"start_date": "2025-01-03",
												"payment_source_id": "ps-001",
												"category_id": "cat-employment",
												"is_active": true,
												"created_at": "2025-01-01T00:00:00Z",
												"updated_at": "2025-01-01T00:00:00Z"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Get all income definitions",
				"summary": "List all incomes",
				"tags": [
					"Incomes"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateIncome",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Income"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Create a new income definition",
				"summary": "Create an income",
				"tags": [
					"Incomes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateIncomeRequest"
							}
						}
					}
				}
			}
		},
		"/api/incomes/{id}": {
			"put": {
				"operationId": "UpdateIncome",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Income"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Update an existing income definition",
				"summary": "Update an income",
				"tags": [
					"Incomes"
				],
				"security": [],
				"parameters": [
					{
						"description": "The income ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateIncomeRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteIncome",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Delete an income definition",
				"summary": "Delete an income",
				"tags": [
					"Incomes"
				],
				"security": [],
				"parameters": [
					{
						"description": "The income ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Category"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "cat-utilities",
												"name": "Utilities",
												"type": "bill",
												"color": "#3b82f6",
												"sort_order": 0,
												"is_predefined": true,
												"created_at": "2025-01-01T00:00:00Z",
												"updated_at": "2025-01-01T00:00:00Z"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Get all categories",
				"summary": "List all categories",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Create a new category",
				"summary": "Create a category",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryRequest"
							}
						}
					}
				}
			}
		},
		"/api/categories/{id}": {
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Update an existing category",
				"summary": "Update a category",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"description": "The category ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCategoryRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Delete a category",
				"summary": "Delete a category",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"description": "The category ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/categories/reorder": {
			"put": {
				"operationId": "ReorderCategories",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Category"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Reorder categories",
				"summary": "Reorder categories",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReorderCategoriesRequest"
							}
						}
					}
				}
			}
		},
		"/api/bills": {
			"get": {
				"operationId": "GetBills",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Bill"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "bill-123",
												"name": "Electric Bill",
												"amount": 15000,
												"billing_period": "monthly",
												"day_of_month": 15,
												"payment_source_id": "ps-001",
												"category_id": "cat-utilities",
												"is_active": true,
												"created_at": "2025-01-01T00:00:00Z",
												"updated_at": "2025-01-01T00:00:00Z"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Get all bill definitions",
				"summary": "List all bills",
				"tags": [
					"Bills"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBill",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Bill"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "bill-123",
											"name": "Electric Bill",
											"amount": 15000,
											"billing_period": "monthly",
											"day_of_month": 15,
											"payment_source_id": "ps-001",
											"category_id": "cat-utilities",
											"is_active": true,
											"created_at": "2025-01-01T00:00:00Z",
											"updated_at": "2025-01-01T00:00:00Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Create a new bill definition",
				"summary": "Create a bill",
				"tags": [
					"Bills"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBillRequest"
							}
						}
					}
				}
			}
		},
		"/api/bills/{id}": {
			"put": {
				"operationId": "UpdateBill",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Bill"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Update an existing bill definition",
				"summary": "Update a bill",
				"tags": [
					"Bills"
				],
				"security": [],
				"parameters": [
					{
						"description": "The bill ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBillRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBill",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Delete a bill definition",
				"summary": "Delete a bill",
				"tags": [
					"Bills"
				],
				"security": [],
				"parameters": [
					{
						"description": "The bill ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}